Rename token renderer to tag renderer
